fix test
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ibm@242499 138bc75d-0d04-0410-961f-82ee72b054a4

+++ fusion/gcc/testsuite/gcc.target/powerpc/fusion4.c
@@ -6,9 +6,8 @@
 
 #define LARGE 0x12345
 
-int fusion_float_read (float *p){ return p[LARGE]; }
+float fusion_float_read (float *p){ return p[LARGE]; }
 
 void fusion_float_write (float *p, float f){ p[LARGE] = f; }
 
-/* { dg-final { scan-assembler "load fusion, type SF"  } } */
 /* { dg-final { scan-assembler "store fusion, type SF" } } */
